+++ lib/k8s/discovery/driver/http.ex
@@ -58,7 +58,7 @@ defmodule K8s.Discovery.Driver.HTTP do
     case K8s.Conn.RequestOptions.generate(conn) do
       {:ok, request_options} ->
         url = Path.join(conn.url, path)
-        headers = conn.http_provider.headers(request_options)
+        headers = K8s.Client.Provider.headers(request_options)
         opts = [ssl: request_options.ssl_options]
         conn.http_provider.request(:get, url, "", headers, opts)
